move sync origin chain id from entityhander to contract entity handler
