+++ src/CAServer.Application.Contracts/UserAssets/IUserAssetsAppService.cs
@@ -19,6 +19,4 @@ public interface IUserAssetsAppService
 
     Task<TokenInfoDto> GetTokenBalanceAsync(GetTokenBalanceRequestDto requestDto);
     
-    Task SyncOriginChainIdAsync(UserLoginEto userLoginEto);
-    Task UpdateOriginChainIdAsync(string originChainId, string syncChainId ,UserLoginEto userLoginEto);
 }
