+++ src/CAServer.Application/CAServerApplicationModule.cs
@@ -51,7 +51,6 @@ public class CAServerApplicationModule : AbpModule
         Configure<ImServerOptions>(configuration.GetSection("ImServer"));
         Configure<HostInfoOptions>(configuration.GetSection("HostInfo"));
         Configure<SeedImageOptions>(configuration.GetSection("SeedSymbolImage"));
-        Configure<SyncOriginChainIdOptions>(configuration.GetSection("SyncOriginChainId"));
         Configure<SecurityOptions>(configuration.GetSection("Security"));
         context.Services.AddSingleton<ISearchService, UserTokenSearchService>();
         context.Services.AddSingleton<ISearchService, ContactSearchService>();
