+++ src/CAServer.ContractEventHandler.Core/Application/IContractAppService.cs
@@ -3,12 +3,17 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
 using System.Threading.Tasks;
+using AElf;
 using AElf.Types;
+using CAServer.Commons;
 using CAServer.Etos;
 using CAServer.Grains.Grain.ApplicationHandler;
+using CAServer.Grains.Grain.ValidateOriginChainId;
 using CAServer.Grains.State.ApplicationHandler;
+using CAServer.Guardian.Provider;
 using CAServer.Monitor;
 using CAServer.Monitor.Logger;
+using CAServer.UserAssets.Provider;
 using CAServer.UserBehavior;
 using CAServer.UserBehavior.Etos;
 using Google.Protobuf;
@@ -18,6 +23,7 @@ using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Newtonsoft.Json;
 using Nito.AsyncEx;
+using Orleans;
 using Portkey.Contracts.CA;
 using Volo.Abp.Caching;
 using Volo.Abp.EventBus.Distributed;
@@ -32,6 +38,9 @@ public interface IContractAppService
     Task QueryAndSyncAsync();
 
     Task InitializeIndexAsync();
+    
+    Task SyncOriginChainIdAsync(UserLoginEto userLoginEto);
+    // Task UpdateOriginChainIdAsync(string originChainId, string syncChainId ,UserLoginEto userLoginEto);
     // Task InitializeQueryRecordIndexAsync();
     // Task InitializeIndexAsync(long blockHeight);
 }
@@ -49,11 +58,18 @@ public class ContractAppService : IContractAppService
     private readonly IIndicatorLogger _indicatorLogger;
     private readonly IMonitorLogProvider _monitorLogProvider;
     private readonly IDistributedCache<string> _distributedCache;
+    private readonly IGuardianProvider _guardianProvider;
+    private readonly IClusterClient _clusterClient;
+    private readonly SyncOriginChainIdOptions _syncOriginChainIdOptions;
+    private readonly IUserAssetsProvider _userAssetsProvider;
 
     public ContractAppService(IDistributedEventBus distributedEventBus, IOptionsSnapshot<ChainOptions> chainOptions,
         IOptionsSnapshot<IndexOptions> indexOptions, IGraphQLProvider graphQLProvider,
         IContractProvider contractProvider, IObjectMapper objectMapper, ILogger<ContractAppService> logger,
         IRecordsBucketContainer recordsBucketContainer, IIndicatorLogger indicatorLogger,
+        IGuardianProvider guardianProvider, IClusterClient clusterClient,
+        IOptions<SyncOriginChainIdOptions> syncOriginChainIdOptions,
+        IUserAssetsProvider userAssetsProvider,
         IMonitorLogProvider monitorLogProvider, IDistributedCache<string> distributedCache)
     {
         _distributedEventBus = distributedEventBus;
@@ -67,6 +83,10 @@ public class ContractAppService : IContractAppService
         _indicatorLogger = indicatorLogger;
         _monitorLogProvider = monitorLogProvider;
         _distributedCache = distributedCache;
+        _guardianProvider = guardianProvider;
+        _clusterClient = clusterClient;
+        _syncOriginChainIdOptions = syncOriginChainIdOptions.Value;
+        _userAssetsProvider = userAssetsProvider;
     }
 
     public async Task CreateHolderInfoAsync(AccountRegisterCreateEto message)
@@ -264,6 +284,125 @@ public class ContractAppService : IContractAppService
         // ValidateAndSync can be very time consuming, so don't wait for it to finish
         _ = ValidateTransactionAndSyncAsync(socialRecoveryDto.ChainId, outputGetHolderInfo, "",
             MonitorTag.SocialRecover);
+    }
+    
+    public async Task SyncOriginChainIdAsync(UserLoginEto userLoginEto)
+    {
+        if (!await NeedSyncStatusAsync(userLoginEto.UserId))
+        {
+            return;
+        }
+
+        var originChainId = "";
+        var syncChainId = "";
+        var guardians = await _guardianProvider.GetGuardiansAsync("", userLoginEto.CaHash);
+        if (guardians == null || guardians.CaHolderInfo == null || guardians.CaHolderInfo.Count == 0)
+        {
+            _logger.LogInformation("CheckOriginChainIdStatusAsync fail,guardians is null or empty,userId {uid}",
+                userLoginEto.UserId);
+            return;
+        }
+
+        originChainId = guardians.CaHolderInfo?.FirstOrDefault()?.OriginChainId;
+        if (string.IsNullOrWhiteSpace(originChainId))
+        {
+            _logger.LogInformation("CheckOriginChainIdStatusAsync fail,originChainId is null or empty,userId {uid}",
+                userLoginEto.UserId);
+            return;
+        }
+
+        syncChainId = _chainOptions.ChainInfos.Where(kvp => kvp.Key != originChainId).Select(kvp => kvp.Key)
+            .FirstOrDefault();
+        if (string.IsNullOrWhiteSpace(syncChainId))
+        {
+            _logger.LogInformation("CheckOriginChainIdStatusAsync fail,syncChainId is null or empty,userId {uid}",
+                userLoginEto.UserId);
+            return;
+        }
+
+        //this will take very long time
+       await UpdateOriginChainIdAsync(originChainId, syncChainId, userLoginEto);
+    }
+
+    public async Task UpdateOriginChainIdAsync(string originChainId, string syncChainId, UserLoginEto userLoginEto)
+    {
+        var validateOriginChainIdGrain = _clusterClient.GetGrain<IValidateOriginChainIdGrain>(userLoginEto.UserId);
+        try
+        {
+            var needValidate = await validateOriginChainIdGrain.NeedValidateAsync();
+            _logger.LogInformation(
+                "UpdateOriginChainIdAsync,needValidate {needValidate},cahash:{cahash},uid:{uid} ,originChainId:{originChainId}",
+                needValidate.Data, userLoginEto.CaHash, userLoginEto.UserId, originChainId);
+
+            if (!needValidate.Data)
+            {
+                return;
+            }
+
+            var holderInfoOutput =
+                await _contractProvider.GetHolderInfoAsync(Hash.LoadFromHex(userLoginEto.CaHash),
+                    null, originChainId);
+
+            var syncHolderInfoOutput =
+                await _contractProvider.GetHolderInfoAsync(Hash.LoadFromHex(userLoginEto.CaHash),
+                    null, syncChainId);
+
+            if (holderInfoOutput.CreateChainId > 0 && syncHolderInfoOutput.CreateChainId > 0)
+            {
+                await validateOriginChainIdGrain.SetStatusSuccessAsync();
+                _logger.LogInformation(
+                    "UpdateOriginChainIdAsync already success,chainId {chainId},userId {uid}", originChainId,
+                    userLoginEto.UserId);
+                return;
+            }
+
+            _logger.LogInformation(
+                "UpdateOriginChainIdAsync success,originChainId {originChainId}:{holderInfoOutput.CreateChainId}, syncChainId:{syncChainId}:{syncHolderInfoOutput.CreateChainId},userId {uid}",
+                originChainId, holderInfoOutput.CreateChainId, syncChainId, syncHolderInfoOutput.CreateChainId,
+                userLoginEto.UserId);
+
+            holderInfoOutput.CreateChainId = ChainHelper.ConvertBase58ToChainId(originChainId);
+
+            var grain = _clusterClient.GetGrain<IContractServiceGrain>(Guid.NewGuid());
+            var transactionDto =
+                await grain.ValidateTransactionAsync(originChainId, holderInfoOutput, null);
+
+            await validateOriginChainIdGrain.SetInfoAsync(transactionDto.TransactionResultDto.TransactionId,
+                originChainId);
+            await validateOriginChainIdGrain.SetStatusSuccessAsync();
+            _ = ValidateTransactionAndSyncAsync(originChainId, holderInfoOutput, "",
+                MonitorTag.LoginSync);
+        }
+        catch (Exception e)
+        {
+            _logger.LogError(e, "UpdateOriginChainIdAsync fail,chainId {chainId},userId {uid},cahash:{cahash}",
+                originChainId,
+                userLoginEto.UserId, userLoginEto.CaHash);
+            await validateOriginChainIdGrain.SetStatusFailAsync();
+        }
+    }
+
+    public async Task<bool> NeedSyncStatusAsync(Guid userId)
+    {
+        var caHolderIndex = await _userAssetsProvider.GetCaHolderIndexAsync(userId);
+        if (caHolderIndex == null || caHolderIndex.IsDeleted)
+        {
+            _logger.LogInformation("UpdateOriginChainIdAsync caHolderIndex is null or deleted,userId {uid}", userId);
+            return false;
+        }
+
+        _logger.LogInformation(
+            "UpdateOriginChainIdAsync caHolderIndex.CreateTime:{caHolderIndex.CreateTime},checkTime:{time}",
+            (TimeHelper.GetTimeStampFromDateTime(caHolderIndex.CreateTime),
+                _syncOriginChainIdOptions.CheckUserRegistrationTimestamp));
+
+        if (TimeHelper.GetTimeStampFromDateTime(caHolderIndex.CreateTime) >
+            _syncOriginChainIdOptions.CheckUserRegistrationTimestamp)
+        {
+            return false;
+        }
+
+        return true;
     }
 
     private async Task ValidateTransactionAndSyncAsync(string chainId, GetHolderInfoOutput result,
