+++ src/CAServer.ContractEventHandler.Core/Application/IContractProvider.cs
@@ -29,7 +29,7 @@ public interface IContractProvider
 {
     Task<GetHolderInfoOutput> GetHolderInfoFromChainAsync(string chainId,
         Hash loginGuardian, string caHash);
-
+    Task<GetHolderInfoOutput> GetHolderInfoAsync(Hash caHash, Hash loginGuardianIdentifierHash, string chainId);
     Task<int> GetChainIdAsync(string chainId);
     Task<long> GetBlockHeightAsync(string chainId);
     Task<TransactionResultDto> GetTransactionResultAsync(string chainId, string txId);
@@ -125,6 +125,18 @@ public class ContractProvider : IContractProvider
         }
     }
 
+    public async Task<GetHolderInfoOutput> GetHolderInfoAsync(Hash caHash, Hash loginGuardianIdentifierHash,
+        string chainId)
+    {
+        var param = new GetHolderInfoInput();
+        param.CaHash = caHash;
+        param.LoginGuardianIdentifierHash = loginGuardianIdentifierHash;
+        var output =
+            await CallTransactionAsync<GetHolderInfoOutput>(chainId, AElfContractMethodName.GetHolderInfo, param,
+                false);
+        return output;
+    }
+    
     public async Task<GetHolderInfoOutput> GetHolderInfoFromChainAsync(string chainId, Hash loginGuardian,
         string caHash)
     {
