+++ src/CAServer.ContractEventHandler.Core/Application/UserLoginHandler.cs
@@ -1,21 +1,22 @@
 using System;
 using System.Threading.Tasks;
+using CAServer.ContractEventHandler.Core.Application;
 using CAServer.Etos;
 using CAServer.UserAssets;
 using Microsoft.Extensions.Logging;
 using Volo.Abp.DependencyInjection;
 using Volo.Abp.EventBus.Distributed;
 
-namespace CAServer.EntityEventHandler.Core;
+namespace CAServer.ContractEventHandler.Core.Application;
 
 public class UserLoginHandler : IDistributedEventHandler<UserLoginEto>,ITransientDependency
 {
-    private readonly IUserAssetsAppService _userAssetsAppService;
+    private readonly IContractAppService _contractAppService;
     private readonly ILogger<UserLoginHandler> _logger;
     
-    public UserLoginHandler(IUserAssetsAppService userAssetsAppService, ILogger<UserLoginHandler> logger)
+    public UserLoginHandler(IContractAppService contractAppService, ILogger<UserLoginHandler> logger)
     {
-        _userAssetsAppService = userAssetsAppService;
+        _contractAppService = contractAppService;
         _logger = logger;
     }
     
@@ -23,7 +24,7 @@ public class UserLoginHandler : IDistributedEventHandler<UserLoginEto>,ITransien
     {
         try 
         {
-            await _userAssetsAppService.SyncOriginChainIdAsync(eventData);
+            await _contractAppService.SyncOriginChainIdAsync(eventData);
         }
         catch (Exception e)
         {
