+++ src/CAServer.ContractEventHandler.Core/CAServerContractEventHandlerCoreModule.cs
@@ -1,5 +1,9 @@
 using CAServer.ContractEventHandler.Core.Application;
+using CAServer.Grains.Grain.ValidateOriginChainId;
 using CAServer.Signature;
+using GraphQL.Client.Abstractions;
+using GraphQL.Client.Http;
+using GraphQL.Client.Serializer.Newtonsoft;
 using Microsoft.Extensions.DependencyInjection;
 using Volo.Abp;
 using Volo.Abp.AutoMapper;
@@ -8,7 +12,7 @@ using Volo.Abp.Threading;
 
 namespace CAServer.ContractEventHandler.Core;
 
-[DependsOn(typeof(AbpAutoMapperModule), typeof(CAServerSignatureModule))]
+[DependsOn(typeof(AbpAutoMapperModule), typeof(CAServerSignatureModule), typeof(CAServerApplicationModule))]
 public class CAServerContractEventHandlerCoreModule : AbpModule
 {
     public override void ConfigureServices(ServiceConfigurationContext context)
@@ -21,6 +25,10 @@ public class CAServerContractEventHandlerCoreModule : AbpModule
 
         var configuration = context.Services.GetConfiguration();
         Configure<CrossChainOptions>(configuration.GetSection("CrossChain"));
+        Configure<SyncOriginChainIdOptions>(configuration.GetSection("SyncOriginChainId"));
+        context.Services.AddSingleton(new GraphQLHttpClient(configuration["GraphQL:Configuration"],
+            new NewtonsoftJsonSerializer()));
+        context.Services.AddScoped<IGraphQLClient>(sp => sp.GetRequiredService<GraphQLHttpClient>());
     }
 
     public override void OnPostApplicationInitialization(ApplicationInitializationContext context)
